Moved CPU class definition to it's own file
