+++ archey/archey.py
@@ -20,6 +20,7 @@ from kernel import Kernel
 from uptime import Uptime
 from disk import Disk
 from ram import RAM
+from cpu import CPU
 from constants import (
     COLOR_DICT,
     DE_DICT,
@@ -238,32 +239,6 @@ class Packages:
             packages = CONFIG.get('default_strings')['not_detected']
 
         self.value = packages
-
-
-class CPU:
-    def __init__(self):
-        model_name_regex = re.compile(
-            r'^model name\s*:\s*(.*)$',
-            flags=re.IGNORECASE | re.MULTILINE
-        )
-
-        with open('/proc/cpuinfo') as file:
-            cpuinfo = re.search(model_name_regex, file.read())
-
-        # This test case has been built for some ARM architectures (see #29).
-        # Sometimes, `model name` info is not present within `/proc/cpuinfo`.
-        # We use the output of `lscpu` program (util-linux-ng) to retrieve it.
-        if not cpuinfo:
-            cpuinfo = re.search(
-                model_name_regex,
-                check_output(
-                    ['lscpu'],
-                    env={'LANG': 'C'}, universal_newlines=True
-                )
-            )
-
-        # Sometimes CPU model name contains extra ugly white-spaces.
-        self.value = re.sub(r'\s+', ' ', cpuinfo.group(1))
 
 
 class GPU:
